Hide the secret messages across image data using image processing.
The main moto behind the project is image processing. where we could able to split with the help of cryptographic algoritham and converted into cypher text and after whatever  text containt should be encrypted in that we are adding that hidden message that perticular file properties after the file property data is built once again the original image we are going to transfer that image from one destination to another destination they also required same software and download image which already hidden message if you apply to this perticular it will decrypt once again text data seperately and image data seperately so this main concept of our project where we are hidding text file and image file this is our main thing.